requirements.txt
dnspython
pydantic
python_dotenv
PyYAML
tabulate